backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
google-genai==0.3.0
pydantic==2.6.3
